feat: system-consolidation-validation-test-file-exemption-burn-down-002 - Preserve workstation outcome-route validation evidence
